Dev updates

+++ code/services/comment-service/src/server.go
@@ -6,6 +6,10 @@ import (
 	dbpb "gen/db-service/pb"
 	models "gen/models/pb"
 	threadpb "gen/thread-service/pb"
+	"math"
+
+	"google.golang.org/grpc/codes"
+	"google.golang.org/grpc/status"
 	"google.golang.org/protobuf/types/known/emptypb"
 )
 
@@ -16,6 +20,18 @@ type CommentServer struct {
 }
 
 func (s *CommentServer) ListComments(ctx context.Context, req *commentpb.ListCommentsRequest) (*commentpb.ListCommentsResponse, error) {
+	// input validation
+	if req.ThreadId != nil && *req.ThreadId == "" {
+		return nil, status.Errorf(codes.InvalidArgument, "thread id cannot be empty")
+	}
+	if req.Offset != nil && *req.Offset < 0 {
+		return nil, status.Errorf(codes.InvalidArgument, "offset must be a non-negative integer")
+	}
+	if req.Limit != nil && *req.Limit <= 0 {
+		return nil, status.Errorf(codes.InvalidArgument, "limit must be a positive integer")
+	}
+
+	// call the database service to fetch comments
 	res, err := s.DBClient.ListComments(ctx, &dbpb.ListCommentsRequest{
 		ThreadId: req.ThreadId,
 		Offset:   req.Offset,
@@ -31,22 +47,18 @@ func (s *CommentServer) ListComments(ctx context.Context, req *commentpb.ListCom
 }
 
 func (s *CommentServer) CreateComment(ctx context.Context, req *commentpb.CreateCommentRequest) (*commentpb.CreateCommentResponse, error) {
-	var err error
-	if req.ParentType == models.CommentParentType_THREAD {
-		// check if thread exists
-		_, err = s.ThreadClient.GetThread(ctx, &threadpb.GetThreadRequest{
-			Id: req.ParentId,
-		})
+	// input validation
+	if req.ParentId == "" {
+		return nil, status.Errorf(codes.InvalidArgument, "parent id is required")
+	}
+	if req.Content == "" {
+		return nil, status.Errorf(codes.InvalidArgument, "content is required")
+	}
+	if len(req.Content) > 500 {
+		return nil, status.Errorf(codes.InvalidArgument, "content exceeds maximum length of 500 characters")
+	}
 
-	} else {
-		// check if comment exists
-		_, err = s.GetComment(ctx, &commentpb.GetCommentRequest{
-			Id: req.ParentId,
-		})
-	}
-	if err != nil {
-		return nil, err
-	}
+	// TODO: update parent num_comments
 
 	// create comment
 	res, err := s.DBClient.CreateComment(ctx, &dbpb.CreateCommentRequest{
@@ -63,32 +75,66 @@ func (s *CommentServer) CreateComment(ctx context.Context, req *commentpb.Create
 }
 
 func (s *CommentServer) GetComment(ctx context.Context, req *commentpb.GetCommentRequest) (*models.Comment, error) {
+	// Input validation
+	if req.Id == "" {
+		return nil, status.Errorf(codes.InvalidArgument, "id is required")
+	}
+
+	// fetch comment
 	res, err := s.DBClient.GetComment(ctx, &dbpb.GetCommentRequest{
 		Id: req.Id,
 	})
 	if err != nil {
 		return nil, err
 	}
+
 	return res, nil
 }
 
 func (s *CommentServer) UpdateComment(ctx context.Context, req *commentpb.UpdateCommentRequest) (*emptypb.Empty, error) {
+	// input validation
+	if req.Id == "" {
+		return nil, status.Errorf(codes.InvalidArgument, "id is required")
+	}
+	if req.Content != nil && len(*req.Content) > 500 {
+		return nil, status.Errorf(codes.InvalidArgument, "content exceeds maximum length of 500 characters")
+	}
+	if req.VoteOffset != nil && math.Abs(float64(*req.VoteOffset)) != 1 {
+		return nil, status.Errorf(codes.InvalidArgument, "invalid votes offset %d", req.VoteOffset)
+	}
+	if req.NumCommentsOffset != nil && math.Abs(float64(*req.NumCommentsOffset)) != 1 {
+		return nil, status.Errorf(codes.InvalidArgument, "invalid num comments offset %d", req.NumCommentsOffset)
+	}
+
+	// update comment
 	_, err := s.DBClient.UpdateComment(ctx, &dbpb.UpdateCommentRequest{
-		Id:      req.Id,
-		Content: req.Content,
+		Id:                req.Id,
+		Content:           req.Content,
+		VoteOffset:        req.VoteOffset,
+		NumCommentsOffset: req.NumCommentsOffset,
 	})
 	if err != nil {
 		return nil, err
 	}
-	return nil, nil
+
+	return &emptypb.Empty{}, nil
 }
 
 func (s *CommentServer) DeleteComment(ctx context.Context, req *commentpb.DeleteCommentRequest) (*emptypb.Empty, error) {
+	// input validation
+	if req.Id == "" {
+		return nil, status.Errorf(codes.InvalidArgument, "id is required")
+	}
+
+	// call the database service to delete a comment
 	_, err := s.DBClient.DeleteComment(ctx, &dbpb.DeleteCommentRequest{
 		Id: req.Id,
 	})
 	if err != nil {
 		return nil, err
 	}
+
+	// TODO: update parent num_comments
+
 	return &emptypb.Empty{}, nil
 }
