+++ code/services/thread-service/src/server.go
@@ -6,6 +6,10 @@ import (
 	dbpb "gen/db-service/pb"
 	models "gen/models/pb"
 	threadpb "gen/thread-service/pb"
+	"math"
+
+	"google.golang.org/grpc/codes"
+	"google.golang.org/grpc/status"
 	"google.golang.org/protobuf/types/known/emptypb"
 )
 
@@ -32,13 +36,23 @@ func (s *ThreadServer) ListThreads(ctx context.Context, req *threadpb.ListThread
 }
 
 func (s *ThreadServer) CreateThread(ctx context.Context, req *threadpb.CreateThreadRequest) (*threadpb.CreateThreadResponse, error) {
-	// check if community exists
-	_, err := s.CommunityClient.GetCommunity(ctx, &communitypb.GetCommunityRequest{
-		Id: req.CommunityId,
-	})
-	if err != nil {
-		return nil, err
+	if req.CommunityId == "" {
+		return nil, status.Errorf(codes.InvalidArgument, "community_id is required")
 	}
+	if req.Title == "" {
+		return nil, status.Errorf(codes.InvalidArgument, "title is required")
+	}
+	if len(req.Title) < 3 || len(req.Title) > 50 {
+		return nil, status.Errorf(codes.InvalidArgument, "title must be between 3 and 50 characters long")
+	}
+	if req.Content == "" {
+		return nil, status.Errorf(codes.InvalidArgument, "content is required")
+	}
+	if len(req.Content) < 3 || len(req.Content) > 500 {
+		return nil, status.Errorf(codes.InvalidArgument, "content must be between 3 and 500 characters long")
+	}
+
+	// TODO: update community num_threads
 
 	// create thread
 	res, err := s.DBClient.CreateThread(ctx, &dbpb.CreateThreadRequest{
@@ -55,6 +69,9 @@ func (s *ThreadServer) CreateThread(ctx context.Context, req *threadpb.CreateThr
 }
 
 func (s *ThreadServer) GetThread(ctx context.Context, req *threadpb.GetThreadRequest) (*models.Thread, error) {
+	if req.Id == "" {
+		return nil, status.Errorf(codes.InvalidArgument, "id is required")
+	}
 	res, err := s.DBClient.GetThread(ctx, &dbpb.GetThreadRequest{
 		Id: req.Id,
 	})
@@ -65,10 +82,27 @@ func (s *ThreadServer) GetThread(ctx context.Context, req *threadpb.GetThreadReq
 }
 
 func (s *ThreadServer) UpdateThread(ctx context.Context, req *threadpb.UpdateThreadRequest) (*emptypb.Empty, error) {
+	if req.Id == "" {
+		return nil, status.Errorf(codes.InvalidArgument, "id is required")
+	}
+	if req.Title != nil && *req.Title != "" && (len(*req.Title) < 3 || len(*req.Title) > 50) {
+		return nil, status.Errorf(codes.InvalidArgument, "title must be between 3 and 50 characters long")
+	}
+	if req.Content != nil && *req.Content != "" && (len(*req.Content) < 3 || len(*req.Content) > 500) {
+		return nil, status.Errorf(codes.InvalidArgument, "content must be between 3 and 500 characters long")
+	}
+	if req.VoteOffset != nil && math.Abs(float64(*req.VoteOffset)) != 1 {
+		return nil, status.Errorf(codes.InvalidArgument, "invalid vote offset %d", req.VoteOffset)
+	}
+	if req.NumCommentsOffset != nil && math.Abs(float64(*req.NumCommentsOffset)) != 1 {
+		return nil, status.Errorf(codes.InvalidArgument, "invalid num comments offset %d", req.NumCommentsOffset)
+	}
 	_, err := s.DBClient.UpdateThread(ctx, &dbpb.UpdateThreadRequest{
-		Id:      req.Id,
-		Title:   req.Title,
-		Content: req.Content,
+		Id:                req.Id,
+		Title:             req.Title,
+		Content:           req.Content,
+		VoteOffset:        req.VoteOffset,
+		NumCommentsOffset: req.NumCommentsOffset,
 	})
 	if err != nil {
 		return nil, err
@@ -77,11 +111,17 @@ func (s *ThreadServer) UpdateThread(ctx context.Context, req *threadpb.UpdateThr
 }
 
 func (s *ThreadServer) DeleteThread(ctx context.Context, req *threadpb.DeleteThreadRequest) (*emptypb.Empty, error) {
+	if req.Id == "" {
+		return nil, status.Errorf(codes.InvalidArgument, "id is required")
+	}
 	_, err := s.DBClient.DeleteThread(ctx, &dbpb.DeleteThreadRequest{
 		Id: req.Id,
 	})
 	if err != nil {
 		return nil, err
 	}
+
+	// TODO: update community num_threads
+
 	return &emptypb.Empty{}, nil
 }
